server/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
scikit-learn
nltk
google.generativeai